Began clearing conversion warnings
Typecasting etc

+++ Source/GUI/TimelineComponent.cpp
@@ -21,11 +21,11 @@ TimelineComponent::TimelineComponent(int64 numberOfClips, int64 mixerOffset)
         TimelineClip* clip = new TimelineClip(clips);
         addAndMakeVisible(clip);
         if(clips == 0)
-            clip->setBounds(pixels, 0, _pixelsPerClip, 20);
+            clip->setBounds((int)pixels, 0, (int)_pixelsPerClip, 20);
         else
         {
             pixels += _pixelsPerClip;
-            clip->setBounds(pixels, 0, _pixelsPerClip, 20);
+            clip->setBounds((int)pixels, 0, (int)_pixelsPerClip, 20);
         }
         _clips.push_back(clip);
     }
